pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "incubator-v3"
version = "1.31.0"
description = "Incubator v3 — Raspberry Pi Zero 2W edition"
requires-python = ">=3.11"
dependencies = [
  # Web framework
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  # Database
  "sqlalchemy>=2.0.0",
  # Validation
  "pydantic>=2.7.0",
  "email-validator>=2.2.0",
  # Templating
  "jinja2>=3.1.0",
  # Image handling (for vision pre-processing)
  "pillow>=10.0.0",
]

[project.optional-dependencies]
# Install on the Pi itself: pip install -e ".[pi]"
pi = [
  # GPIO — Pi Zero 2W hardware access
  "RPi.GPIO>=0.7.0",
  # DHT22 temperature/humidity sensor
  "adafruit-circuitpython-dht>=3.7.0",
  # Required by adafruit-dht for board pin mapping
  "adafruit-blinka>=8.0.0",
]

# Pi camera (install via apt: sudo apt install python3-picamera2)
# picamera2 is a system package on Raspberry Pi OS — install with:
#   sudo apt install -y python3-picamera2
# Or as fallback for USB cameras: pip install opencv-python-headless

# TFLite inference (optional, for local on-device vision model)
tflite = [
  # Pi-compatible build — install from piwheels:
  #   pip install tflite-runtime --extra-index-url https://www.piwheels.org/simple
  "tflite-runtime>=2.14.0",
  "numpy>=1.24.0",
]

# Full dev environment (no hardware needed — all services run in mock mode)
dev = [
  "httpx>=0.27.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
]

[tool.setuptools]
packages = ["app", "database"]

[tool.setuptools.package-dir]
"" = "."

[tool.uvicorn]
factory = false
host = "0.0.0.0"
port = 8000
